requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
sqlalchemy==2.0.10
asyncmy==0.2.9
pymysql==1.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
pydantic==1.10.7
python-dotenv==1.0.0
